refactor(amazonq): remove duplicate and make trimSupplementalContexts() a util
