+++ server/aws-lsp-codewhisperer/src/shared/supplementalContextUtil/supplementalContextUtil.ts
@@ -2,7 +2,7 @@
 // https://github.com/aws/aws-toolkit-vscode/blob/9d8ddbd85f4533e539a58e76f7c46883d8e50a79/packages/core/src/codewhisperer/util/supplementalContext/supplementalContextUtil.ts
 
 import { fetchSupplementalContextForSrc } from './crossFileContextUtil'
-import { CodeWhispererSupplementalContext } from '../models/model'
+import { CodeWhispererSupplementalContext, CodeWhispererSupplementalContextItem } from '../models/model'
 import {
     CancellationToken,
     Logging,
@@ -151,6 +151,57 @@ export function truncateSupplementalContext(
         supplementalContextItems: c,
         contentsLength: curTotalLength,
     }
+}
+
+// Constants for supplemental context limits
+const supplementalContextMaxTotalLength: number = 8192
+const charactersLimit: number = 10000
+
+// TODO: what's the difference between this implementation vs. [truncateSupplementalContext] above?
+/**
+ * Trims the supplementalContexts array to ensure it doesn't exceed the max number
+ * of contexts or total character length limit
+ *
+ * @param supplementalContextItems - Array of CodeWhispererSupplementalContextItem objects (already sorted with newest first)
+ * @param maxContexts - Maximum number of supplemental contexts allowed
+ * @returns Trimmed array of CodeWhispererSupplementalContextItem objects
+ */
+export function trimSupplementalContexts(
+    supplementalContextItems: CodeWhispererSupplementalContextItem[],
+    maxContexts: number
+): CodeWhispererSupplementalContextItem[] {
+    if (supplementalContextItems.length === 0) {
+        return supplementalContextItems
+    }
+
+    // First filter out any individual context that exceeds the character limit
+    let result = supplementalContextItems.filter(context => {
+        return context.content.length <= charactersLimit
+    })
+
+    // Then limit by max number of contexts
+    if (result.length > maxContexts) {
+        result = result.slice(0, maxContexts)
+    }
+
+    // Lastly enforce total character limit
+    let totalLength = 0
+    let i = 0
+
+    while (i < result.length) {
+        totalLength += result[i].content.length
+        if (totalLength > supplementalContextMaxTotalLength) {
+            break
+        }
+        i++
+    }
+
+    if (i === result.length) {
+        return result
+    }
+
+    const trimmedContexts = result.slice(0, i)
+    return trimmedContexts
 }
 
 export function truncateLineByLine(input: string, l: number): string {
